Support for URI references to resources
I've modified to code to allow a Resource to supply a uri_prefix.  If it
does, then references to those resources will no longer be simple
object ids, but completely URIs constructed from the uri_prefix and
the object id.

There is one remaining issue in this code which is that I haven't handled
all cases where related_resources is used.