Fixed a bug that caused tests to fail.
I had one lingering issue which showed up when I ran the nose tests.
That issue is fixed and now all the original tests that were written
for the flask_mongorest package pass in the context of my changes.

My one remaining task is to write some additional tests that test
my additions.