Added a test for new POST behavior and fixed on last bug
I had one last case where I had a variable name wrong.  I also modified
the example application to introduce a uri_prefix for one of the resources
and then tested that a successful post to create that resource
returned the resource location in the Location header of the
response.  Furthermore, I made sure that the resulting URL contained
the uri_prefix associated with that resource.