So close...
There appears to be an issue in the tests where a URL for a resource is
passed into the API and the JSON response that is returned from the
POST request shows that the URL for the field has changed.  It looks like
perhaps a **copy** of the resource referenced via the URL is being made.