Added a utility method that generates URIs for a given Resource
One issue I have found is that the normal convention for REST interfaces
is that a URL ending in "/" normally represents a collection of resources
whereas a URL ending with no "/" represents an individual resource.

In terms of passing URIs for resources, we need to be consistent and I've
followed the convention above.

The problem is that the test cases don't really honor this convention.
As a result, there are many requests formulated in the test cases that
would simply lead to a 301 (redirect) status code.  So I set the
`app.url_map.strict_slashes` flag to be false which seems to cause
werkzeug to avoid the redirect and just respond directly to the request.

All of these changes are backward compatible with respect to any code
that pre-dates the existence of the `uri_prefix` in `Resource`.