Added some authorization methods
These methods map exactly to the way Django does things.  The only thing
we are really missing is the "read" permission.  However, that part is
tricky because it interacts with the QuerySet objects.  For that reason,
I left it out (upt to this point) to focus on the ones that were pretty
straightforward.

This includes some extensive tests for these permissions checks.