Added support for read permissions.
The read permission is different than the other permissions because
it has to hook into the QuerySet system.  I implemented this in such
a way that the `has_read_permission` method is given a QuerySet and
returns a QuerySet of objects that the given request should have
read access to.  The default behavior is, of course, to just return
the initial QuerySet.

I added a test for this behavior as well.